Move `stringify_cmd` to `util` and use it everywhere
Also move PythonSdist to starforge/packaging/setup.py to prevent circular imports.
