+++ starforge/cache.py
@@ -19,7 +19,11 @@ from pkg_resources import parse_version
 from six import with_metaclass
 
 from .io import warn, info, debug, fatal
-from .util import pip_install, py_to_pip
+from .util import (
+    pip_install,
+    py_to_pip,
+    stringify_cmd
+)
 
 
 class BaseCacher(with_metaclass(ABCMeta, object)):
@@ -174,7 +178,7 @@ class PipSourceCacher(TarballCacher):
                     '--no-deps', name + '==' + version
                 ]
                 info('Fetching sdist: %s', name)
-                debug('Executing: %s', ' '.join(cmd))
+                debug('Executing: %s', stringify_cmd(cmd))
                 subprocess.check_call(cmd, stdout=sys.stderr)
                 cfpath = self.check(name, version=version)
             except subprocess.CalledProcessError:
@@ -230,7 +234,7 @@ def cache_wheel_sources(cache_manager, wheel_config):
     if wheel_config.setup_requires:
         # this is done due to pypa/pip#1884 - `pip download` fails under certain circumstances if setup_requires is set
         info("Installing packages to Starforge Python at '%s' for '%s' setup requirements: %s",
-            sys.executable, wheel_config.name, ', '.join(wheel_config.setup_requires))
+             sys.executable, wheel_config.name, ', '.join(wheel_config.setup_requires))
         pip_install(pip=py_to_pip(sys.executable), packages=wheel_config.setup_requires)
     sources.append(cache_manager.pip_cache(wheel_config.name, wheel_config.version, fail_ok=fail_ok))
     for src_url in wheel_config.sources:
